@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:piwigo_ng/components/modals/open_tag_modal.dart';
 import 'package:piwigo_ng/components/notification_dot.dart';
 import 'package:piwigo_ng/components/popup_list_item.dart';
 import 'package:piwigo_ng/services/preferences_service.dart';
@@ -40,9 +41,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
       if (widget.scrollController.offset > _expandedHeight * _opacityScale) {
         return 0.0;
       }
-      return (_expandedHeight * _opacityScale -
-              widget.scrollController.offset) /
-          (_expandedHeight * _opacityScale);
+      return (_expandedHeight * _opacityScale - widget.scrollController.offset) / (_expandedHeight * _opacityScale);
     }
     return 1.0;
   }
@@ -58,8 +57,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
       }
 
       // In case 0%-100% of the expanded height is viewed
-      double scrollDelta =
-          (_expandedHeight - widget.scrollController.offset) / _expandedHeight;
+      double scrollDelta = (_expandedHeight - widget.scrollController.offset) / _expandedHeight;
       double scrollPercent = (scrollDelta * 2 - 1);
       return (1 - scrollPercent) * delta * basePadding + basePadding;
     }
@@ -71,8 +69,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
   Widget build(BuildContext context) {
     return SliverAppBar(
       leading: IconButton(
-        onPressed: () =>
-            Navigator.of(context).pushNamed(SettingsPage.routeName),
+        onPressed: () => Navigator.of(context).pushNamed(SettingsPage.routeName),
         icon: const Icon(Icons.settings),
       ),
       pinned: true,
@@ -90,7 +87,6 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
                 child: AppField(
                   padding: const EdgeInsets.symmetric(vertical: 8.0),
                   prefix: Icon(Icons.search),
-                  hint: "Search...",
                 ),
               ),
             ),
@@ -108,7 +104,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
         ),
         title: Text(
           appStrings.tabBar_albums,
-          textScaleFactor: 1,
+          textScaler: TextScaler.linear(1),
           style: Theme.of(context).appBarTheme.titleTextStyle,
         ),
       ),
@@ -125,8 +121,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
             PopupMenuItem(
               onTap: () => Future.delayed(
                 const Duration(seconds: 0),
-                () =>
-                    Navigator.of(context).pushNamed(UploadStatusPage.routeName),
+                () => Navigator.of(context).pushNamed(UploadStatusPage.routeName),
               ),
               child: Stack(
                 children: [
@@ -137,8 +132,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
                   Positioned(
                     top: 14.0,
                     left: 0.0,
-                    child: Consumer<UploadNotifier>(
-                        builder: (context, uploadNotifier, child) {
+                    child: Consumer<UploadNotifier>(builder: (context, uploadNotifier, child) {
                       return NotificationDot(
                         isShown: uploadNotifier.uploadList.isNotEmpty,
                       );
@@ -147,12 +141,21 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
                 ],
               ),
             ),
+            PopupMenuItem(
+              onTap: () => Future.delayed(
+                const Duration(seconds: 0),
+                () => showOpenTagModal(context),
+              ),
+              child: PopupListItem(
+                icon: Icons.local_offer_outlined,
+                text: appStrings.tags,
+              ),
+            ),
             if (Preferences.getUserStatus != 'guest')
               PopupMenuItem(
                 onTap: () => Future.delayed(
                   const Duration(seconds: 0),
-                  () => Navigator.of(context)
-                      .pushNamed(ImageFavoritesPage.routeName),
+                  () => Navigator.of(context).pushNamed(ImageFavoritesPage.routeName),
                 ),
                 child: PopupListItem(
                   icon: Icons.favorite,
@@ -164,8 +167,7 @@ class _RootSearchAppBarState extends State<RootSearchAppBar> {
         Positioned(
           top: 12.0,
           left: 12.0,
-          child: Consumer<UploadNotifier>(
-              builder: (context, uploadNotifier, child) {
+          child: Consumer<UploadNotifier>(builder: (context, uploadNotifier, child) {
             return NotificationDot(
               isShown: uploadNotifier.uploadList.isNotEmpty,
             );
