@@ -33,8 +33,7 @@ class _SettingsAppBarState extends State<SettingsAppBar> {
       }
 
       // In case 0%-100% of the expanded height is viewed
-      double scrollDelta =
-          (_expandedHeight - widget.scrollController.offset) / _expandedHeight;
+      double scrollDelta = (_expandedHeight - widget.scrollController.offset) / _expandedHeight;
       double scrollPercent = (scrollDelta * 2 - 1);
       return (1 - scrollPercent) * delta * basePadding + basePadding;
     }
@@ -59,7 +58,7 @@ class _SettingsAppBarState extends State<SettingsAppBar> {
         ),
         title: Text(
           appStrings.tabBar_preferences,
-          textScaleFactor: 1,
+          textScaler: TextScaler.linear(1),
           style: Theme.of(context).appBarTheme.titleTextStyle,
         ),
       ),
