@@ -1,5 +1,5 @@
 import 'dart:io';
-import 'dart:ui' as ui show Codec, FrameInfo, Image;
+import 'dart:ui' as ui show Image;
 
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:extended_text/extended_text.dart';
@@ -17,8 +17,7 @@ import 'package:provider/provider.dart';
 import 'package:video_player/video_player.dart';
 
 class ImageDetailsCard extends StatelessWidget {
-  const ImageDetailsCard({Key? key, required this.image, this.onRemove})
-      : super(key: key);
+  const ImageDetailsCard({Key? key, required this.image, this.onRemove}) : super(key: key);
 
   final ImageModel image;
   final Function()? onRemove;
@@ -57,9 +56,7 @@ class ImageDetailsCard extends StatelessWidget {
           child: ClipRRect(
             borderRadius: BorderRadius.circular(5.0),
             child: Builder(builder: (context) {
-              final String? imageUrl = image
-                  .getDerivativeFromString(Preferences.getImageThumbnailSize)
-                  ?.url;
+              final String? imageUrl = image.getDerivativeFromString(Preferences.getImageThumbnailSize)?.url;
               return ImageNetworkDisplay(
                 imageUrl: imageUrl,
               );
@@ -91,10 +88,7 @@ class ImageDetailsCard extends StatelessWidget {
               children: [
                 Flexible(
                   child: Text(
-                    image.file
-                        .replaceAll('', '\u200B')
-                        .split(path.extension(image.file))
-                        .first,
+                    image.file.replaceAll('', '\u200B').split(path.extension(image.file)).first,
                     maxLines: 1,
                     overflow: TextOverflow.ellipsis,
                     style: Theme.of(context).textTheme.bodySmall,
@@ -112,13 +106,11 @@ class ImageDetailsCard extends StatelessWidget {
           const Spacer(),
           if (image.dateAvailable != null)
             Builder(builder: (context) {
-              LocaleNotifier localeNotifier =
-                  Provider.of<LocaleNotifier>(context, listen: false);
+              LocaleNotifier localeNotifier = Provider.of<LocaleNotifier>(context, listen: false);
               String date =
-                  DateFormat.yMMMMd(localeNotifier.locale.languageCode)
-                      .format(DateTime.parse(image.dateAvailable!));
-              String time = DateFormat.Hms(localeNotifier.locale.languageCode)
-                  .format(DateTime.parse(image.dateAvailable!));
+                  DateFormat.yMMMMd(localeNotifier.locale.languageCode).format(DateTime.parse(image.dateAvailable!));
+              String time =
+                  DateFormat.Hms(localeNotifier.locale.languageCode).format(DateTime.parse(image.dateAvailable!));
               return AutoSizeText(
                 "$date $time",
                 maxLines: 1,
@@ -150,8 +142,7 @@ class ImageDetailsCard extends StatelessWidget {
 }
 
 class LocalImageDetailsCard extends StatefulWidget {
-  const LocalImageDetailsCard(
-      {Key? key, required this.image, this.onRemove, this.isDuplicate = false})
+  const LocalImageDetailsCard({Key? key, required this.image, this.onRemove, this.isDuplicate = false})
       : super(key: key);
 
   final File image;
@@ -214,17 +205,12 @@ class _LocalImageDetailsCardState extends State<LocalImageDetailsCard> {
             fit: StackFit.expand,
             children: [
               LayoutBuilder(builder: (context, constraints) {
-                List<String>? mimeType =
-                    mime(widget.image.path.split('/').last)?.split('/');
+                List<String>? mimeType = mime(widget.image.path.split('/').last)?.split('/');
 
                 if (mimeType?.first == 'image') {
                   _checkMemory();
-                  double? cacheWidth = constraints.maxWidth.isInfinite
-                      ? constraints.maxWidth
-                      : null;
-                  double? cacheHeight = constraints.maxHeight.isInfinite
-                      ? constraints.maxHeight
-                      : null;
+                  double? cacheWidth = constraints.maxWidth.isInfinite ? constraints.maxWidth : null;
+                  double? cacheHeight = constraints.maxHeight.isInfinite ? constraints.maxHeight : null;
                   return Image.file(
                     widget.image,
                     fit: BoxFit.cover,
@@ -327,8 +313,7 @@ class _LocalImageDetailsCardState extends State<LocalImageDetailsCard> {
 }
 
 class LocalVideoDetailsCard extends StatefulWidget {
-  const LocalVideoDetailsCard(
-      {Key? key, required this.video, this.onRemove, this.isDuplicate = false})
+  const LocalVideoDetailsCard({Key? key, required this.video, this.onRemove, this.isDuplicate = false})
       : super(key: key);
 
   final File video;
@@ -361,9 +346,7 @@ class _LocalVideoDetailsCardState extends State<LocalVideoDetailsCard> {
     final Duration duration = _controller.value.duration;
     int hours = duration.inHours;
     int minutes = (duration - Duration(hours: hours)).inMinutes;
-    int seconds =
-        (duration - Duration(hours: hours) - Duration(minutes: minutes))
-            .inSeconds;
+    int seconds = (duration - Duration(hours: hours) - Duration(minutes: minutes)).inSeconds;
     return '${hours > 0 ? '$hours:' : ''}${minutes < 10 ? '0$minutes' : '$minutes'}:${seconds < 10 ? '0$seconds' : '$seconds'}';
   }
 
@@ -433,17 +416,12 @@ class _LocalVideoDetailsCardState extends State<LocalVideoDetailsCard> {
                       bottom: 2.0,
                       left: 2.0,
                       child: Container(
-                        padding: const EdgeInsets.symmetric(
-                            horizontal: 4, vertical: 2),
+                        padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 2),
                         decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(5),
-                            color: AppColors.black.withOpacity(0.7)),
+                            borderRadius: BorderRadius.circular(5), color: AppColors.black.withOpacity(0.7)),
                         child: Text(
                           _duration,
-                          style: TextStyle(
-                              color: AppColors.white,
-                              fontSize: 10,
-                              fontWeight: FontWeight.bold),
+                          style: TextStyle(color: AppColors.white, fontSize: 10, fontWeight: FontWeight.bold),
                         ),
                       ),
                     ),
